+++ src/main/java/org/geowe/client/local/layermanager/tool/export/exporter/github/GitHubEventListener.java
@@ -1,0 +1,8 @@
+package org.geowe.client.local.layermanager.tool.export.exporter.github;
+
+
+
+
+public interface GitHubEventListener<T> {
+	public void onFinish(T response);	
+}